+++ packages/connect/CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v0.x.x - _TBD, 2017_
 
-    * Sanitize api endpoint url and remove trailing slashes (#315)
-    * Improve error message text in HttpClient (#315)
+    * Sanitize api endpoint url and remove trailing slashes (#318)
+    * Improve error message text in HttpClient (#318)
+    * Stop appending '/v0' to api endpoint url in HttpClient (#318)
 
 ## v0.4.0 - _January 11, 2017_