+++ packages/connect/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## v0.x.x - _TBD, 2017_
+
+    * Sanitize api endpoint url and remove trailing slashes (#318)
+    * Improve error message text in HttpClient (#318)
+    * Stop appending '/v0' to api endpoint url in HttpClient (#318)
+
 ## v0.4.0 - _January 11, 2017_
 
     * Prevent getFeesAsync method on HttpClient from mutating input (#296)